Remove old wizard with new one
