+++ src/smart-components/group/add-group/schema.js
