+++ src/smart-components/group/add-group/set-roles.js
@@ -1,38 +1,35 @@
-import React, { Fragment } from 'react';
+import React, { Fragment, useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
-import { Card, Form, FormGroup, Stack, StackItem, Text, TextContent, Title } from '@patternfly/react-core';
+import useFieldApi from '@data-driven-forms/react-form-renderer/dist/esm/use-field-api';
+import useFormApi from '@data-driven-forms/react-form-renderer/dist/esm/use-form-api';
+import { FormGroup, Stack, StackItem, Text, TextContent } from '@patternfly/react-core';
 import RolesList from './roles-list';
 import '../../../App.scss';
 
-const SetRoles = ({ selectedRoles, setSelectedRoles, title, description }) => {
+const SetRoles = (props) => {
+  const [selectedRoles, setSelectedRoles] = useState([]);
+  const { input } = useFieldApi(props);
+  const formOptions = useFormApi();
+
+  useEffect(() => {
+    input.onChange(selectedRoles);
+    formOptions.change('selected-roles', selectedRoles);
+  }, [selectedRoles]);
+
   return (
     <Fragment>
-      <Form>
-        <Stack hasGutter>
-          {title && (
-            <StackItem>
-              <Title headingLevel="h4" size="xl">
-                {title}
-              </Title>
-            </StackItem>
-          )}
-          <StackItem>
-            <TextContent>
-              <Title headingLevel="h4" size="xl">
-                Add roles
-              </Title>
-              <Text>{description || 'Select one or more roles to add to this group.'}</Text>
-            </TextContent>
-          </StackItem>
-          <StackItem>
-            <FormGroup fieldId="select-role">
-              <Card>
-                <RolesList selectedRoles={selectedRoles} setSelectedRoles={setSelectedRoles} />
-              </Card>
-            </FormGroup>
-          </StackItem>
-        </Stack>
-      </Form>
+      <Stack hasGutter>
+        <StackItem>
+          <TextContent>
+            <Text>Select one or more roles to add to this group.</Text>
+          </TextContent>
+        </StackItem>
+        <StackItem>
+          <FormGroup fieldId="select-role">
+            <RolesList selectedRoles={selectedRoles} setSelectedRoles={setSelectedRoles} />
+          </FormGroup>
+        </StackItem>
+      </Stack>
     </Fragment>
   );
 };
