+++ src/smart-components/group/add-group/set-users.js
@@ -1,28 +1,29 @@
-import React, { Fragment } from 'react';
+import React, { Fragment, useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
-import { Card, Form, FormGroup, Stack, StackItem, TextContent, Title } from '@patternfly/react-core';
+import { Card, Form, FormGroup, Stack, StackItem, TextContent } from '@patternfly/react-core';
+import useFieldApi from '@data-driven-forms/react-form-renderer/dist/esm/use-field-api';
+import useFormApi from '@data-driven-forms/react-form-renderer/dist/esm/use-form-api';
 import { CompactUsersList } from './users-list';
 import ActiveUser from '../../../presentational-components/shared/ActiveUsers';
 import '../../../App.scss';
 
-const SetUsers = ({ selectedUsers, setSelectedUsers, title, description }) => {
+const SetUsers = (props) => {
+  const [selectedUsers, setSelectedUsers] = useState([]);
+  const { input } = useFieldApi(props);
+  const formOptions = useFormApi();
+
+  useEffect(() => {
+    input.onChange(selectedUsers);
+    formOptions.change('selected-users', selectedUsers);
+  }, [selectedUsers]);
+
   return (
     <Fragment>
       <Form>
         <Stack hasGutter>
-          {title && (
-            <StackItem>
-              <Title headingLevel="h4" size="xl">
-                {title}
-              </Title>
-            </StackItem>
-          )}
           <StackItem>
             <TextContent>
-              <Title headingLevel="h4" size="xl">
-                Add members
-              </Title>
-              <ActiveUser description={description || 'These are all of the users in your Red Hat organization. To manage users, go to your'} />
+              <ActiveUser description={'These are all of the users in your Red Hat organization. To manage users, go to your'} />
             </TextContent>
           </StackItem>
           <StackItem>
